refactor: split CLI into focused modules and add shared retry utility
CLI split (1615 -> 659 lines in __init__.py):
- cli/output.py — Colors, formatting helpers, output utils
- cli/cmd_init.py — agentos init command
- cli/cmd_validate.py — agentos validate with JSON Schema support
- cli/cmd_audit.py — agentos audit with CSV export
- cli/cmd_policy.py — agentos policy dispatcher
- cli/policy_checker.py — PolicyChecker, PolicyViolation, rule loading

All public symbols re-exported from cli/__init__.py for backward
compatibility. Existing tests pass unchanged (67 passed, 8 skipped).

Shared retry utility (retry.py):
- Decorator for sync + async functions with exponential backoff
- Configurable max_attempts, backoff_base, exception types
- on_retry callback support
- 13 tests
